+++ lua/lazyvim/plugins/extras/coding/native_snippets.lua
@@ -11,18 +11,23 @@ return {
   },
   {
     "nvim-cmp",
-    opts = {
-      snippet = {
+    dependencies = {
+      { "rafamadriz/friendly-snippets" },
+      { "garymjr/nvim-snippets", opts = { friendly_snippets = true } },
+    },
+    opts = function(_, opts)
+      opts.snippet = {
         expand = function(args)
           vim.snippet.expand(args.body)
         end,
-      },
-    },
+      }
+      table.insert(opts.sources, { name = "snippets" })
+    end,
     keys = {
       {
         "<Tab>",
         function()
-          if vim.snippet.jumpable(1) then
+          if vim.snippet.active({ direction = 1 }) then
             vim.schedule(function()
               vim.snippet.jump(1)
             end)
@@ -47,7 +52,7 @@ return {
       {
         "<S-Tab>",
         function()
-          if vim.snippet.jumpable(-1) then
+          if vim.snippet.active({ direction = -1 }) then
             vim.schedule(function()
               vim.snippet.jump(-1)
             end)