+++ lua/lazyvim/plugins/extras/lang/python.lua
@@ -2,9 +2,11 @@ if lazyvim_docs then
   -- LSP Server to use for Python.
   -- Set to "basedpyright" to use basedpyright instead of pyright.
   vim.g.lazyvim_python_lsp = "pyright"
+  vim.g.lazyvim_python_ruff = "ruff_lsp"
 end
 
 local lsp = vim.g.lazyvim_python_lsp or "pyright"
+local ruff = vim.g.lazyvim_python_ruff or "ruff_lsp"
 
 return {
   {
@@ -29,6 +31,12 @@ return {
           enabled = true,
         },
         ruff_lsp = {
+          enabled = ruff == "ruff_lsp",
+        },
+        ruff = {
+          enabled = ruff == "ruff",
+        },
+        [ruff] = {
           keys = {
             {
               "<leader>co",
@@ -47,9 +55,9 @@ return {
         },
       },
       setup = {
-        ruff_lsp = function()
+        [ruff] = function()
           LazyVim.lsp.on_attach(function(client, _)
-            if client.name == "ruff_lsp" then
+            if client.name == ruff then
               -- Disable hover in favor of Pyright
               client.server_capabilities.hoverProvider = false
             end