+++ lua/lazyvim/plugins/extras/lang/typescript.lua
@@ -2,6 +2,17 @@ local debugtypePreferred = {
   ['node'] = 'pwa-node',
 }
 
+local inlay_hints_settings = {
+  includeInlayEnumMemberValueHints = true,
+  includeInlayFunctionLikeReturnTypeHints = true,
+  includeInlayFunctionParameterTypeHints = true,
+  includeInlayParameterNameHints = "literal",
+  includeInlayParameterNameHintsWhenArgumentMatchesName = false,
+  includeInlayPropertyDeclarationTypeHints = true,
+  includeInlayVariableTypeHints = false,
+  includeInlayVariableTypeHintsWhenTypeMatchesName = false,
+}
+
 return {
 
   -- add typescript to treesitter
@@ -50,8 +61,13 @@ return {
               desc = "Remove Unused Imports",
             },
           },
-          ---@diagnostic disable-next-line: missing-fields
           settings = {
+            typescript = {
+              inlayHints = inlay_hints_settings,
+            },
+            javascript = {
+              inlayHints = inlay_hints_settings,
+            },
             completions = {
               completeFunctionCalls = true,
             },