+++ lua/lazyvim/util/toggle.lua
@@ -67,7 +67,7 @@ function M.inlay_hints(buf, value)
     if value == nil then
       value = not ih.is_enabled(buf)
     end
-    ih.enable(buf, value)
+    ih.enable(value, { bufnr = buf })
   end
 end